ingestion/service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
sentence-transformers==2.7.0
numpy==1.26.4
